[BSE-1797] Support object_keys (#6604)
Adds support for the Snowflake JSON function OBJECT_KEYS which takes in a JSON column and returns a column where each row is an array of the string keys of that row. This is our first real JSON function in BodoSQL, so some extra testing infrastructure setup (and BdooSQL context logic) was necessary to add. Logic to toggle between map vs struct arrays was ported from this PR: #6398

+++ bodo/libs/map_arr_ext.py
@@ -655,8 +655,8 @@ def map_arr_getitem(arr, ind):
             offsets = bodo.libs.array_item_arr_ext.get_offsets(arr._data)
             struct_arr = bodo.libs.array_item_arr_ext.get_data(arr._data)
             key_data, value_data = bodo.libs.struct_arr_ext.get_data(struct_arr)
-            start_offset = offsets[ind]
-            end_offset = offsets[ind + 1]
+            start_offset = offsets[np.int64(ind)]
+            end_offset = offsets[np.int64(ind) + 1]
             for i in range(start_offset, end_offset):
                 out[key_data[i]] = value_data[i]
             return out
